requirements.txt
torch==1.5.1
torchvision==0.6.1
opencv-python==4.3.0.36
numpy==1.19.0
scipy==1.5.1
pyyaml==5.3.1
numba==0.50.1
filterpy==1.4.5